refactor: 예외처리를 좀 더 올바른 곳에서 행하고 그에 맞게 테스트 코드 수정
